+++ jest.setup.ts
@@ -12,8 +12,8 @@ global.chrome = {
   },
   storage: {
     local: {
-      get: jest.fn((keys, callback) => callback({})),
-      set: jest.fn((items, callback) => callback())
+      get: jest.fn((_keys, callback) => callback({})),
+      set: jest.fn((_items, callback) => callback())
     }
   }
 } as unknown as typeof chrome;